app/requirements.txt
streamlit>=1.31.0

# Core LangChain
langchain>=0.2.10
langchain-core>=0.2.10
langchain-community>=0.2.10

# Integrations
langchain-groq>=0.1.5
langchain-huggingface>=0.0.3

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector Store (Windows-safe)
faiss-cpu>=1.7.4

# PDF Loading
pypdf>=4.0.0

# Utilities
python-dotenv>=1.0.1
